requirements.txt
# Core dependencies
flask==2.0.1
flask-login==0.6.2
flask-sqlalchemy==3.0.3
google-generativeai==0.3.1
pandas==1.5.3
python-dotenv==1.0.0
gunicorn==20.1.0

# Data handling and export
openpyxl==3.1.2
sqlalchemy==2.0.20
pyarrow==12.0.1
fastparquet==2023.7.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# API and integration
requests==2.30.0
beautifulsoup4==4.12.2
gspread==5.10.0
oauth2client==4.1.3
pymysql==1.1.0
psycopg2-binary==2.9.6

# Multi-threading and async
celery==5.3.1
redis==4.6.0

# Utility
tqdm==4.65.0
pyyaml==6.0.1
faker==19.3.0
